--- src/app.js
@@ -1,31 +1,0 @@
-const express = require("express");
-const posts_router = require("./routes/posts_router");
-const comments_router = require("./routes/comments_router");
-const app = express();
-
-const swaggerOptions = {
-    swaggerDefinition: {
-        openapi: '3.0.0',
-        info: {
-            title: 'Comments API',
-            version: '1.0.0',
-            description: 'API for managing comments',
-        },
-        servers: [
-            {
-                url: 'http://localhost:3000',
-            },
-        ],
-    },
-    apis: ['./**/*.ts'],
-};
-
-const swaggerDocs = swaggerJsDoc(swaggerOptions);
-app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocs));
-
-app.use(express.json());
-
-app.use("/posts", posts_router);
-app.use("/comments", comments_router);
-
-module.exports = app;
