+++ src/middlewares/auth.ts
@@ -1,0 +1,30 @@
+import { Request, Response, NextFunction } from 'express';
+import jwt, { JwtPayload } from 'jsonwebtoken';
+
+interface AuthenticatedRequest extends Request {
+  user?: string | JwtPayload;
+}
+
+export const authenticateToken = (
+  req: AuthenticatedRequest,
+  res: Response,
+  next: NextFunction
+): void => {
+  const token = req.headers['authorization']?.split(' ')[1];
+
+  if (!token) {
+    res.status(401).json({ error: 'Access token required.' });
+    return;
+  }
+
+  jwt.verify(token, 'secretKey', (err, user) => {
+    if (err) {
+      res.status(403).json({ error: 'Invalid token.' });
+      return;
+    }
+
+    req.user = user;
+    next();
+  });
+};
+
